requirements.txt
skillsnetwork
numpy
pandas
Pillow
scikit-image
torch
torchvision
gradio
